Add files via upload
The project utilizes the Google API to retrieve comprehensive data from YouTube channels. The data includes information on channels, playlists, videos, and comments. By interacting with the Google API, we collect the data and merge it into a JSON file. YouTube Data Harvesting and Warehousing is a project that aims to allow users to access and analyze data from multiple YouTube channels.